merges filter comp with page
